+++ CHANGELOG.md
@@ -9,6 +9,16 @@
 
 * `New-BBServerSession` function to replace `New-BBServerConnection`.
 * `Session` parameter on all functions to replace `Connection` parameter.
+* `Disable-BBServerRepository` for archinving repositories.
+* `Enable-BBServerRepository` for unarchiving repositories.
+
+### Changed
+
+* `Invoke-BBServerRestMethod` now shows request URL and request body in PowerShell's verbose stream.
+
+### Fixed
+
+`Invoke-BBServerRestMethod` would fail if a response didn't return JSON, which happens with some bad requests.
 
 ### Deprecated